requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.13.1
annotated-types==0.7.0
anyio==4.12.1
asn1crypto==1.5.1
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloud-sql-python-connector==1.20.0
cryptography==46.0.5
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.110.0
frozenlist==1.8.0
google-auth==2.48.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.0
idna==3.11
iniconfig==2.3.0
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.1
packaging==26.0
pg8000==1.31.5
pluggy==1.6.0
propcache==0.4.1
psycopg2-binary==2.9.9
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.0
pydantic_core==2.41.5
pytest==8.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
scramp==1.4.8
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.25
starlette==0.36.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.29.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
yarl==1.22.0
